pyproject.toml
[tool.poetry]
name = "mlops"
version = "0.1.0"
description = ""
authors = ["Ainur Zagitov <ainur175@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
urllib3 = "1.26.15"
tqdm = "^4.66.1"
matplotlib = "^3.8.2"
torch = "^2.1.2"
torchvision = "^0.16.2"
torch-summary = "^1.4.5"
pydantic = "1.10.7"
python-dotenv = "1.0.0"
scikit-learn = "^1.3.2"
pre-commit = "^3.6.0"
dvc = "^3.38.1"
dvc-gdrive = "^3.0.1"
mlflow = "^2.9.2"
typing-extensions = "4.2.0"
psycopg2-binary = "^2.9.9"
hydra-core = "^1.3.2"
fire = "^0.5.0"

[tool.black]
line-length = 90
target-version = ["py310"]

[tool.flake8]
max-line-length = 110

[tool.isort]
src_paths = ["ds_project"]
profile = "black"
line_length = 90
lines_after_imports = 2


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
